requirements.txt
fastapi~=0.63.0
uvicorn~=0.13.0
SQLAlchemy~=1.3.22
pymysql==0.9.3
alembic~=1.4.3
pydantic~=1.7.3
redis==3.5.3
python-multipart==0.0.5
aiofiles==0.5.0
loguru~=0.5.3
gunicorn==20.0.4
cryptography==3.1
starlette~=0.13.6
PyJWT~=1.7.1